% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{adivmeasures}
\alias{adivmeasures}
\title{Calculating Alpha Diversity Measures Per Sample}
\usage{
adivmeasures(abundance_data)
}
\arguments{
\item{abundance_data}{data frame, containing ASV/OTU abundance data per sample}
}
\value{
a data frame containing each alpha diversity measure for each sample
}
\description{
This function calculates alpha diversity measures for each sample. These
measures includes species richness, number of individuals, species evenness,
Shannon Diversity Index, and inverse Simpson Diversity index. A csv file of
the resulting data frame is also written to the working directory.
}
\examples{
NA
}
