% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{batch_correct}
\alias{batch_correct}
\title{Adjusting for Batch Effect}
\usage{
batch_correct(feature_csv, metadata_csv)
}
\arguments{
\item{feature_csv}{a csv file that contains raw ASV/OTU counts of each sample}

\item{metadata_csv}{a csv file that contains metadata for each sample; there MUST be a 'Batch' variable present}
}
\value{
a list of data frames: original feature and metadata input files along with the adjusted ASV/OTU table; a csv file of the adjusted ASV/OTU table is written to the working directory
}
\description{
This function allows the user to correct their ASV/OTU abundance data due to
the presence of the "batch effect" in their data.  Adjusting for any
batch effect allows for the user's analyses to be statistically sound against
any introduced bias from any introduced batches in their data. If the batch
effect is NOT present in the data, then batch correction is NOT required.
}
\examples{
NA
}
