% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{batch_test}
\alias{batch_test}
\title{Testing for Batch Effect}
\usage{
batch_test(abundance_data, metadata)
}
\arguments{
\item{abundance_data}{data frame, containing ASV/OTU abundance data per sample}

\item{metadata}{string, csv file that contains metadata for each sample; there MUST be a 'Batch' variable present}
}
\value{
ANOSIM results (parametric testing) or PERMANOVA results (non-parametric testing)
}
\description{
This function allows the user to test their ASV/OTU abundance data for the
"batch effect". In order to perform this test, the metadata input file MUST
contain a 'Batch' column/variable.
}
\examples{
NA
}
