% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{betadiv_stats}
\alias{betadiv_stats}
\title{Creating Bray-Curtis Distance Matrix & Running Beta Diversity Stats}
\usage{
betadiv_stats(abundance_data, metadata_var)
}
\arguments{
\item{abundance_data}{data frame, containing ASV/OTU abundance data per sample}

\item{metadata_var}{a variable within the metadata data frame (format: metadata$Variable)}
}
\value{
a list containing the resulting distance matrix and statistical analysis outcomes
}
\description{
This function allows the user to create a distance matrix using Bray-Curtis
distances. Once the distance matrix is created, statistical analyses are
performed using the resulting distance matrix and desired variable.
}
\examples{
NA
}
