% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{calculate_abund.metadata}
\alias{calculate_abund.metadata}
\title{Create Filtered Abundance Tables using Raw OTU/ASV input (Illumina MiSeq)}
\usage{
calculate_abund.metadata(feature_csv, metadata_csv)
}
\arguments{
\item{feature_csv}{string, csv file that contains raw ASV/OTU counts of each sample}

\item{metadata_csv}{string, csv file that contains metadata for each sample}
}
\value{
a list of data frames that are each different abundance filters which can be used for further analyses
}
\description{
This function allows the user to create numerous ASV/OTU abundance tables
(ASV/OTUs occurring less than 0.1%, ASV/OTUs occurring less than 0.01%,
dominant ASV/OTUs, presence/absence form of dominant ASV/OTUs, and relative
abundances of ASV/OTUs occurring less than 0.1%) each of which can
perform further microbiome analyses and visualization dependent on the project.
Metadata input is used to correspond samples to their respective metadata;
any samples without metadata is NOT included within the resulting data frames.
}
\examples{
NA
}
