% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{calculate_abund}
\alias{calculate_abund}
\title{Create Filtered Abundance Tables using Raw OTU/ASV input (Illumina MiSeq)}
\usage{
calculate_abund(feature_csv)
}
\arguments{
\item{feature_csv}{a csv file that contains raw ASV/OTU counts of each sample}
}
\value{
a list of data frames that are each different abundance filters
}
\description{
This function allows the user to create numerous ASV/OTU abundance tables
(ASV/OTUs occurring less than 0.1%, ASV/OTUs occurring less than 0.01%,
dominant ASV/OTUs, presence/absence form of dominant ASV/OTUs, and relative
abundances of ASV/OTUs occurring less than 0.1%) each of which can
perform further microbiome analyses and visualization dependent on the project.
Metadata input is NOT used within this function, thus diversity analyses
CANNOT be performed if metadata is required.
}
\examples{
NA
}
