% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{create_phyloseq}
\alias{create_phyloseq}
\title{Creating a Phyloseq Object using Phyloseq}
\usage{
create_phyloseq(abund, taxonomy, metadata)
}
\arguments{
\item{abund}{a dataframe containing abundance information; created using calculate_abund.metadata or calculate_abund functions}

\item{taxonomy}{a csv file that contains taxonomic information, ASV/OTU should be rows while the taxonomic information should be columns}

\item{metadata}{a csv file that contains any metadata on your samples, samples should be rows}
}
\value{
a phyloseq object in which the abundances have been transformed to relative abundances
}
\description{
This function allows the user to create a phyloseq object in order to
perform further taxonomic analyses and visualization using the
package 'phyloseq'.
}
\examples{
NA
}
