% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{nonp_kruskal}
\alias{nonp_kruskal}
\title{Non-parametric Testing for Alpha Diversity}
\usage{
nonp_kruskal(adiv_var, test_var)
}
\arguments{
\item{adiv_var}{a variable within the metadata data frame that accounts for an alpha diversity measure (example: metadata$inv.D)}

\item{test_var}{a variable within the metadata data frame (format: metadata$Variable)}
}
\value{
a list containing the results of the Kruskal-Wallis test, pairwise Wilcoxon test, and the letters showing the significance between each variable level
}
\description{
This function is used to perform a Kruskal-Wallis and pairwise Wilcoxon test
to analyze the statistical significance of an alpha diversity measure against
a given variable.
}
\examples{
NA
}
