% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{p_anova}
\alias{p_anova}
\title{Parametric Testing for Alpha Diversity}
\usage{
p_anova(adiv_var, test_var)
}
\arguments{
\item{adiv_var}{a variable within the metadata data frame that accounts for an alpha diversity measure (example: metadata$inv.D)}

\item{test_var}{a variable within the metadata data frame (format: metadata$Variable)}
}
\value{
a list containing an ANOVA table and results of the Tukey test
}
\description{
This function is used to perform an ANOVA and pairwise Tukey test to analyze
the statistical significance of an alpha diversity measure against a given
variable.
}
\examples{
NA
}
