% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{select.envfit}
\alias{select.envfit}
\title{Filtering envfit vectors (vegan)}
\usage{
select.envfit(fit, r.select)
}
\arguments{
\item{fit}{result of ennfit function in vegan}

\item{r.select}{numeric, correlation minimum threshold}
}
\value{
fits as a result of the function selections
}
\description{
Filters CCA/nMDS/MDS envfit vectors based on r² values. Additionally, this
allows the user to display only significant values in their final plot.
Beckers, Niklas. (2017). Re: How do I set cutoff r² values for plotting
arrows from function envfit in R?. Retrieved from:
https://www.researchgate.net/post/How_do_I_set_cutoff_r_values_for_plotting_arrows_from_function_envfit_in_R/59ba8e1c96b7e411171a6b8c/citation/download.
}
\examples{
NA
}
