% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{setwd_seed}
\alias{setwd_seed}
\title{Set Working Directory and Seed}
\usage{
setwd_seed(file_path, seed_number)
}
\arguments{
\item{file_path}{a string input, a file path to the folder where all generated files will be stored}

\item{seed_number}{a random number that is user-defined}
}
\value{
This function does not return anything
}
\description{
This function allows the user to set their working directory and seed using
only one function. This is a required step when performing any kind of
analyses in RStudio to allow for reproducible results.
}
\examples{
setwd_seed("C:/Users/Public/Documents",2023)
}
