% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{taxon_aggregate}
\alias{taxon_aggregate}
\title{Aggregation on Taxonomic Levels}
\usage{
taxon_aggregate(phyloseq_object, tax_level)
}
\arguments{
\item{phyloseq_object}{phyloseq object, containing taxonomic information}

\item{tax_level}{string, desired taxonomic level}
}
\value{
a data frame that includes count data of the indicated taxonomic level
}
\description{
This function is used to aggregate taxonomic data on a desired taxonomic level
and display the overall abundance of the indicated taxonomic level. A csv file
of the resulting count data frame is also written to the working directory.
}
\examples{
NA
}
