% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBiome16S_functions.R
\name{top_taxa}
\alias{top_taxa}
\title{Calculating Top Taxa}
\usage{
top_taxa(phyloseq_object, tax_level, n)
}
\arguments{
\item{phyloseq_object}{phyloseq object, containing taxonomic information}

\item{tax_level}{string, desired taxonomic level}

\item{n}{numeric, number of desired top taxa}
}
\value{
a phyloseq object that is filtered to the desired number of top taxa
}
\description{
This function calculates the top taxa (user-defined) and their abundances. A
csv file of the top taxa counts are written to the working directory.
}
\examples{
NA
}
